Fix: FE bugs
