+++ ui/client/stylesheets/processes.styl
@@ -95,7 +95,7 @@
       width: 40%
     &.reactable-th-modifydate, &.reactable-th-createdat, &.reactable-th-isdeployed
       width: 10%
-    &.reactable-th-status, &.reactable-th-edit, &.reactable-th-metrics, &.reactable-th-subprocess
+    &.reactable-th-status, &.reactable-th-edit, &.reactable-th-metrics, &.reactable-th-subprocess, &.reactable-th-cancel
       width: 5%
     &:focus
       outline 2px solid focusColor
@@ -126,7 +126,7 @@
     padding-bottom: 5px
   }
 
-  .centered-column, .reactable-th-modifydate, .reactable-th-createdat, .reactable-th-subprocess, .reactable-th-createdby
+  .centered-column, .reactable-th-modifydate, .reactable-th-createdat, .reactable-th-subprocess, .reactable-th-createdby, .reactable-th-cancel
     text-align: center
 
   .edit-column, .reactable-th-metrics, .reactable-th-view
